requirements.txt
Appium-Python-Client
pytest
selenium
pytest-rerunfailures
allure-pytest
behave
allure-behave
pytest-xdist
pytest-randomly
pytest-parametrized
requests